[Refactor] Move `validateRayClusterStatus` function to `validation.go`  and move unit test to  `validation_test.go` (#2780)
