+++ ray-operator/controllers/ray/raycluster_controller_unit_test.go
@@ -3832,82 +3832,3 @@ func TestValidateRayClusterSpecSuspendingWorkerGroup(t *testing.T) {
 		})
 	}
 }
-
-func TestValidateRayClusterStatus(t *testing.T) {
-	tests := []struct {
-		name        string
-		conditions  []metav1.Condition
-		expectError bool
-	}{
-		{
-			name: "Both suspending and suspended are true",
-			conditions: []metav1.Condition{
-				{
-					Type:   string(rayv1.RayClusterSuspending),
-					Status: metav1.ConditionTrue,
-				},
-				{
-					Type:   string(rayv1.RayClusterSuspended),
-					Status: metav1.ConditionTrue,
-				},
-			},
-			expectError: true,
-		},
-		{
-			name: "Only suspending is true",
-			conditions: []metav1.Condition{
-				{
-					Type:   string(rayv1.RayClusterSuspending),
-					Status: metav1.ConditionTrue,
-				},
-				{
-					Type:   string(rayv1.RayClusterSuspended),
-					Status: metav1.ConditionFalse,
-				},
-			},
-			expectError: false,
-		},
-		{
-			name: "Only suspended is true",
-			conditions: []metav1.Condition{
-				{
-					Type:   string(rayv1.RayClusterSuspending),
-					Status: metav1.ConditionFalse,
-				},
-				{
-					Type:   string(rayv1.RayClusterSuspended),
-					Status: metav1.ConditionTrue,
-				},
-			},
-			expectError: false,
-		},
-		{
-			name: "Both suspending and suspended are false",
-			conditions: []metav1.Condition{
-				{
-					Type:   string(rayv1.RayClusterSuspending),
-					Status: metav1.ConditionFalse,
-				},
-				{
-					Type:   string(rayv1.RayClusterSuspended),
-					Status: metav1.ConditionFalse,
-				},
-			},
-			expectError: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			instance := &rayv1.RayCluster{
-				Status: rayv1.RayClusterStatus{
-					Conditions: tt.conditions,
-				},
-			}
-			err := validateRayClusterStatus(instance)
-			if (err != nil) != tt.expectError {
-				t.Errorf("validateRayClusterStatus() error = %v, wantErr %v", err, tt.expectError)
-			}
-		})
-	}
-}
